requirements.txt
Django==2.1.1
django-tables2==1.21.2
factory-boy==2.11.1